Add terraform batch  load spot blueprint
Downloaded from:
https://${REGION}.console.aws.amazon.com/ec2sp/v2/home?region=${REGION}#/spot/blueprints

This quick start template of a batch workload using Spot Instances inside an AWS Batch managed
compute environment. When you choose to configure the Spot blueprint, you can make selections
to generate a configuration that meets your application requirements. Spot blueprint allows you to
configure a managed compute environment and the required dependencies such as IAM roles and
network, and attaches a job queue to it. After you complete all of the steps to configure the Spot
blueprint, you can download it as CloudFormation or Terraform formats, or deploy using the AWS
CloudFormation console.